+++ hpc/nvidia/cuda/cuda_adapter.h
@@ -1719,3 +1719,12 @@
  * NCCL: NVIDIA Collective Communication Library
  */
 #define ncclGetVersion(version)	__nccl(GetVersion(version))
+
+/**
+ * bfloat16
+ */
+#if defined(__USE_HPCC__)
+# define __nv_bfloat16	__hpcc_bfloat16
+#elif defined(__USE_HIP__)
+# define __nv_bfloat16	__hip_bfloat16
+#endif