+++ hpc/nvidia/cuda/intrinsics/bfloat16.cu
@@ -1,0 +1,25 @@
+// SPDX-License-Identifier: GPL-3.0
+/* Copyright (c) 2025 Rong Tao */
+/**
+ * https://docs.nvidia.com/cuda/cuda-math-api/index.html
+ * https://docs.nvidia.com/cuda/cuda-math-api/cuda_math_api/group__CUDA__MATH__INTRINSIC__BFLOAT16.html
+ */
+#include <assert.h>
+#include <stdio.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <stdbool.h>
+#include <math.h>
+#include <string.h>
+#include "cuda_compat.h"
+#include "compiler.h"
+#include "print.h"
+
+
+int main(int argc, char *argv[])
+{
+	assert(sizeof(__nv_bfloat16) == 2 && "bad size of half");
+
+	(void)cudaDeviceSynchronize();
+	return 0;
+}